+++ harborapi/models/base.py
@@ -9,9 +9,10 @@ for more information: <https://rich.readthedocs.io/en/latest/protocol.html#conso
 from __future__ import annotations
 
 from typing import Any
-from typing import Dict
 from typing import Iterable
+from typing import Iterator
 from typing import Optional
+from typing import Sequence
 from typing import Set
 from typing import Type
 from typing import TypeVar
@@ -54,25 +55,22 @@ T = TypeVar("T")
 class RootModel(PydanticRootModel[T]):
     model_config = ConfigDict(validate_assignment=True)
 
+    root: T
+
     def __bool__(self) -> bool:
         return bool(self.root)
 
-
-class StrDictRootModel(RootModel[Optional[Dict[str, T]]]):
-    # All JSON keys are string, so the key type does need to be
-    # parameterized with a generic type.
-
-    def __iter__(self) -> Any:
+    def __iter__(self) -> Iterator[Any]:
         # TODO: fix API spec so root  types can never be none, only
         # the empty container. That way we can always iterate and access
         # without checking for None.
-        if self.root is not None:
-            return iter(self.root)
+        if isinstance(self.root, Iterable):
+            return iter(self.root)  # pyright: ignore[reportUnknownArgumentType, reportUnknownMemberType, reportUnknownVariableType]
         return iter([])
 
-    def __getitem__(self, item: str) -> Optional[T]:
-        if self.root is not None:
-            return self.root[item]
+    def __getitem__(self, key: Any) -> Any:
+        if isinstance(self.root, Sequence):
+            return self.root[key]
         return None
 
     # Enables dot access to dict keys for backwards compatibility
@@ -81,11 +79,6 @@ class StrDictRootModel(RootModel[Optional[Dict[str, T]]]):
             return self.root[attr]  # type: ignore # forego None check and let KeyError raise
         except (KeyError, TypeError):
             raise AttributeError(f"{self.__class__.__name__} has no attribute {attr}")
-
-
-class StrRootModel(RootModel[str]):
-    def __str__(self) -> str:
-        return str(self.root)
 
 
 class BaseModel(PydanticBaseModel):